+++ terminal/input/input.go
@@ -14,6 +14,7 @@ package input
 import (
 	"fmt"
 	"os"
+	"strconv"
 	"strings"
 	"unicode/utf8"
 
@@ -34,6 +35,9 @@ type CompletionHandler = func(input string) []string
 
 // HintHandler is hint handler
 type HintHandler = func(input string) string
+
+// Validator is input validation function
+type Validator = func(input string) (string, error)
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -71,13 +75,54 @@ var NewLine = false
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+var (
+	// NotEmpty returns an error if input is empty
+	NotEmpty = validatorNotEmpty
+
+	// IsNumber returns an error if the input is not a valid number
+	IsNumber = validatorIsNumber
+
+	// IsFloat returns an error if the input is not a valid floating number
+	IsFloat = validatorIsFloat
+
+	// IsEmail returns an error if the input is not a valid email
+	IsEmail = validatorIsEmail
+
+	// IsURL returns an error if the input is not a valid URL
+	IsURL = validatorIsURL
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+var (
+	// ErrInvalidAnswer is error for wrong answer for Y/N question
+	ErrInvalidAnswer = fmt.Errorf("Please enter Y or N")
+
+	// ErrIsEmpty is error for empty input
+	ErrIsEmpty = fmt.Errorf("You must enter non-empty value")
+
+	// ErrInvalidNumber is error for invalid number
+	ErrInvalidNumber = fmt.Errorf("Entered value is not a valid number")
+
+	// ErrInvalidFloat is error for invalid floating number
+	ErrInvalidFloat = fmt.Errorf("Entered value is not a valid floating number")
+
+	// ErrInvalidEmail is error for invalid email
+	ErrInvalidEmail = fmt.Errorf("Entered value is not a valid e-mail")
+
+	// ErrInvalidURL is error for invalid URL
+	ErrInvalidURL = fmt.Errorf("Entered value is not a valid URL")
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 var oldTMUXFlag int8
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // Read reads user input
-func Read(title string, nonEmpty bool) (string, error) {
-	return readUserInput(title, nonEmpty, false)
+func Read(title string, validators ...Validator) (string, error) {
+	return readUserInput(title, false, validators)
 }
 
 // ReadAnswer reads user's answer to yes/no question
@@ -104,7 +149,7 @@ func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
 
 	for {
 		answer, err := readUserInput(
-			getAnswerTitle(title, defaultAnswer), false, false,
+			getAnswerTitle(title, defaultAnswer), false, nil,
 		)
 
 		if err != nil {
@@ -121,7 +166,7 @@ func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
 		case "N":
 			return false, nil
 		default:
-			terminal.Warn("Please enter Y or N")
+			terminal.Warn(ErrInvalidAnswer)
 			fmtc.NewLine()
 		}
 	}
@@ -129,14 +174,14 @@ func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
 
 // ReadPassword reads password or some private input that will be hidden
 // after pressing Enter
-func ReadPassword(title string, nonEmpty bool) (string, error) {
-	return readUserInput(title, nonEmpty, true)
+func ReadPassword(title string, validators ...Validator) (string, error) {
+	return readUserInput(title, true, validators)
 }
 
 // ReadPasswordSecure reads password or some private input that will be hidden
 // after pressing Enter
-func ReadPasswordSecure(title string, nonEmpty bool) (*secstr.String, error) {
-	password, err := readUserInput(title, nonEmpty, true)
+func ReadPasswordSecure(title string, validators ...Validator) (*secstr.String, error) {
+	password, err := readUserInput(title, true, validators)
 
 	if err != nil {
 		return nil, err
@@ -163,6 +208,93 @@ func SetCompletionHandler(h CompletionHandler) {
 // SetHintHandler adds function for input hints
 func SetHintHandler(h HintHandler) {
 	linenoise.SetHintHandler(h)
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// validatorNotEmpty is validator for empty input
+func validatorNotEmpty(input string) (string, error) {
+	if strings.TrimSpace(input) != "" {
+		return input, nil
+	}
+
+	return input, ErrIsEmpty
+}
+
+// validatorIsNumber is validator for number
+func validatorIsNumber(input string) (string, error) {
+	input = strings.TrimSpace(input)
+
+	if input == "" {
+		return input, nil // Empty imput is okay
+	}
+
+	_, err := strconv.ParseInt(input, 10, 64)
+
+	if err != nil {
+		return input, ErrInvalidNumber
+	}
+
+	return input, nil
+}
+
+// validatorIsFloat is validator for floating number
+func validatorIsFloat(input string) (string, error) {
+	input = strings.TrimSpace(input)
+
+	if input == "" {
+		return input, nil // Empty imput is okay
+	}
+
+	_, err := strconv.ParseFloat(input, 64)
+
+	if err != nil {
+		return input, ErrInvalidFloat
+	}
+
+	return input, nil
+}
+
+// validatorIsEmail is validator for email
+func validatorIsEmail(input string) (string, error) {
+	input = strings.TrimSpace(input)
+
+	if input == "" {
+		return input, nil // Empty imput is okay
+	}
+
+	name, domain, ok := strings.Cut(input, "@")
+
+	if !ok || strings.TrimSpace(name) == "" ||
+		strings.TrimSpace(domain) == "" || !strings.Contains(domain, ".") {
+		return input, ErrInvalidEmail
+	}
+
+	return input, nil
+}
+
+// validatorIsURL is validator for URL
+func validatorIsURL(input string) (string, error) {
+	input = strings.TrimSpace(input)
+
+	if input == "" {
+		return input, nil // Empty imput is okay
+	}
+
+	switch {
+	case strings.HasPrefix(input, "http://"),
+		strings.HasPrefix(input, "https://"),
+		strings.HasPrefix(input, "ftp://"):
+		// OK
+	default:
+		return input, ErrInvalidURL
+	}
+
+	if !strings.Contains(input, ".") {
+		return input, ErrInvalidURL
+	}
+
+	return input, nil
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -211,7 +343,7 @@ func getAnswerTitle(title, defaultAnswer string) string {
 }
 
 // readUserInput reads user input
-func readUserInput(title string, nonEmpty, private bool) (string, error) {
+func readUserInput(title string, private bool, validators []Validator) (string, error) {
 	if title != "" {
 		fmtc.Println(TitleColorTag + title + "{!}")
 	}
@@ -223,6 +355,7 @@ func readUserInput(title string, nonEmpty, private bool) (string, error) {
 		linenoise.SetMaskMode(true)
 	}
 
+INPUT_LOOP:
 	for {
 		input, err = linenoise.Line(fmtc.Sprintf(Prompt))
 
@@ -234,9 +367,17 @@ func readUserInput(title string, nonEmpty, private bool) (string, error) {
 			return "", err
 		}
 
-		if nonEmpty && strings.TrimSpace(input) == "" {
-			terminal.Warn("\nYou must enter non-empty value\n")
-			continue
+		if len(validators) != 0 {
+			for _, validator := range validators {
+				input, err = validator(input)
+
+				if err != nil {
+					fmtc.NewLine()
+					terminal.Warn(err.Error())
+					fmtc.NewLine()
+					continue INPUT_LOOP
+				}
+			}
 		}
 
 		if private && input != "" {