+++ terminal/input/input_stubs.go
@@ -19,8 +19,21 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+// CompletionHandler is completion handler
+type CompletionHandler = func(input string) []string
+
+// HintHandler is hint handler
+type HintHandler = func(input string) string
+
+// Validator is input validation function
+type Validator = func(input string) (string, error)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // ❗ ErrKillSignal is error type when user cancel input
 var ErrKillSignal = errors.New("")
+
+// ////////////////////////////////////////////////////////////////////////////////// //
 
 // ❗ Prompt is prompt string
 var Prompt = "> "
@@ -51,8 +64,49 @@ var NewLine = false
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+var (
+	// NotEmpty returns an error if input is empty
+	NotEmpty = func(input string) (string, error) { return "", nil }
+
+	// IsNumber returns an error if the input is not a valid number
+	IsNumber = func(input string) (string, error) { return "", nil }
+
+	// IsFloat returns an error if the input is not a valid floating number
+	IsFloat = func(input string) (string, error) { return "", nil }
+
+	// IsEmail returns an error if the input is not a valid email
+	IsEmail = func(input string) (string, error) { return "", nil }
+
+	// IsURL returns an error if the input is not a valid URL
+	IsURL = func(input string) (string, error) { return "", nil }
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+var (
+	// ErrInvalidAnswer is error for wrong answer for Y/N question
+	ErrInvalidAnswer = errors.New("")
+
+	// ErrIsEmpty is error for empty input
+	ErrIsEmpty = errors.New("")
+
+	// ErrInvalidNumber is error for invalid number
+	ErrInvalidNumber = errors.New("")
+
+	// ErrInvalidFloat is error for invalid floating number
+	ErrInvalidFloat = errors.New("")
+
+	// ErrInvalidEmail is error for invalid email
+	ErrInvalidEmail = errors.New("")
+
+	// ErrInvalidURL is error for invalid URL
+	ErrInvalidURL = errors.New("")
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // ❗ Read reads user input
-func Read(title string, nonEmpty bool) (string, error) {
+func Read(title string, validators ...Validator) (string, error) {
 	panic("UNSUPPORTED")
 }
 
@@ -63,13 +117,13 @@ func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
 
 // ❗ ReadPassword reads password or some private input that will be hidden
 // after pressing Enter
-func ReadPassword(title string, nonEmpty bool) (string, error) {
+func ReadPassword(title string, validators ...Validator) (string, error) {
 	panic("UNSUPPORTED")
 }
 
 // ❗ ReadPasswordSecure reads password or some private input that will be hidden
 // after pressing Enter
-func ReadPasswordSecure(title string, nonEmpty bool) (*secstr.String, error) {
+func ReadPasswordSecure(title string, validators ...Validator) (*secstr.String, error) {
 	panic("UNSUPPORTED")
 }