+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Changelog
 
-### [13.9.3](https://kaos.sh/ek/13.9.3)
+### [13.10.0](https://kaos.sh/ek/13.10.0)
+
+> [!CAUTION]
+> This release contains breaking changes to the `input.Read`, `input.ReadPassword`, and `input.ReadPasswordSecure` methods. Prior to this release, all of these methods took a boolean argument to disallow empty input. Since we are adding input validators, you will need to use the `NotEmpty` validator for the same behaviour.
 
 - `[fmtutil/table]` Added automatic breaks feature
+- `[terminal/input]` Added input validation feature
 
 ### [13.9.2](https://kaos.sh/ek/13.9.2)